backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via
    #   -r requirements.in
    #   unstructured-client
aiohappyeyeballs==2.5.0
    # via aiohttp
aiohttp==3.11.13
    # via langchain-community
aiologger==0.7.0
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
alembic==1.15.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   litestar
    #   watchfiles
attrs==25.1.0
    # via
    #   aiohttp
    #   environ-config
authlib==1.3.1
    # via weaviate-client
backoff==2.2.1
    # via unstructured
beautifulsoup4==4.13.3
    # via unstructured
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via unstructured
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   litestar
    #   nltk
    #   rich-click
    #   uvicorn
cryptography==44.0.2
    # via
    #   authlib
    #   unstructured-client
dataclasses-json==0.6.7
    # via
    #   langchain-community
    #   unstructured
editorconfig==0.17.0
    # via jsbeautifier
emoji==2.14.1
    # via unstructured
environ-config==24.1.0
    # via -r requirements.in
eval-type-backport==0.2.2
    # via unstructured-client
faker==37.0.0
    # via polyfactory
fast-query-parsers==1.0.3
    # via litestar
filetype==1.2.0
    # via unstructured
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.70.0
    # via
    #   grpcio-health-checking
    #   grpcio-tools
    #   weaviate-client
grpcio-health-checking==1.70.0
    # via weaviate-client
grpcio-tools==1.70.0
    # via weaviate-client
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   langsmith
    #   litestar
    #   unstructured-client
    #   weaviate-client
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via litestar
joblib==1.4.2
    # via nltk
jsbeautifier==1.15.4
    # via litestar
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.20
    # via langchain-community
langchain-community==0.3.19
    # via -r requirements.in
langchain-core==0.3.43
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.3.6
    # via langchain
langdetect==1.0.9
    # via unstructured
langsmith==0.3.13
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
litestar[standard]==2.15.1
    # via -r requirements.in
litestar-htmx==0.4.1
    # via litestar
lxml==5.3.1
    # via unstructured
mako==1.3.9
    # via alembic
markdown==3.7
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
marshmallow==3.26.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
msgspec==0.19.0
    # via litestar
multidict==6.1.0
    # via
    #   aiohttp
    #   litestar
    #   yarl
multipart==1.2.1
    # via litestar
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via unstructured-client
nltk==3.9.1
    # via unstructured
numpy==2.0.2
    # via
    #   langchain-community
    #   unstructured
orjson==3.10.15
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
polyfactory==2.19.0
    # via litestar
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
    # via
    #   grpcio-health-checking
    #   grpcio-tools
psycopg2==2.9.10
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
    #   unstructured-client
    #   weaviate-client
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via langchain-community
pygments==2.19.1
    # via rich
pypdf==5.3.1
    # via unstructured-client
python-dateutil==2.9.0.post0
    # via unstructured-client
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-iso639==2025.2.18
    # via unstructured
python-magic==0.4.27
    # via unstructured
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   litestar
    #   uvicorn
rapidfuzz==3.12.2
    # via unstructured
regex==2024.11.6
    # via nltk
requests==2.32.3
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   unstructured
requests-toolbelt==1.0.0
    # via
    #   langsmith
    #   unstructured-client
rich==13.9.4
    # via
    #   litestar
    #   rich-click
rich-click==1.8.6
    # via litestar
six==1.17.0
    # via
    #   jsbeautifier
    #   langdetect
    #   python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.38
    # via
    #   alembic
    #   langchain
    #   langchain-community
tabulate==0.9.0
    # via unstructured
tenacity==9.0.0
    # via
    #   langchain-community
    #   langchain-core
tqdm==4.67.1
    # via
    #   nltk
    #   unstructured
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   beautifulsoup4
    #   langchain-core
    #   litestar
    #   polyfactory
    #   pydantic
    #   pydantic-core
    #   rich-click
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
    #   unstructured
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.0
    # via unstructured-client
tzdata==2025.1
    # via faker
unstructured==0.14.8
    # via -r requirements.in
unstructured-client==0.31.1
    # via unstructured
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via litestar
uvloop==0.21.0
    # via
    #   litestar
    #   uvicorn
validators==0.34.0
    # via weaviate-client
watchfiles==1.0.4
    # via uvicorn
weaviate-client==4.11.1
    # via -r requirements.in
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via unstructured
yarl==1.18.3
    # via aiohttp
zstandard==0.23.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# setuptools
